Fix warnings

+++ Source/Tools/relive_api/TlvObjectBaseAOBase.cpp
@@ -16,14 +16,13 @@
 TlvObjectBaseAOBase::TlvObjectBaseAOBase(std::size_t sizeOfT, AO::TlvTypes tlvType, const std::string& typeName, AO::Path_TLV* pSelfTlv)
     : TlvObjectBase(typeName), mSizeOfT(sizeOfT), mType(tlvType), mPSelfTlv{pSelfTlv}
 {
-
 }
 
 TlvObjectBaseAOBase::TlvObjectBaseAOBase(std::size_t sizeOfT, TypesCollectionBase& globalTypes, AO::TlvTypes tlvType, const std::string& typeName, AO::Path_TLV* pSelfTlv, AO::Path_TLV* pTlv, CopyFn copyFn)
     : TlvObjectBase(typeName), mSizeOfT(sizeOfT), mType(tlvType), mPSelfTlv{pSelfTlv}
 {
     mPSelfTlv->field_4_type.mType = mType;
-    mPSelfTlv->field_2_length = mSizeOfT;
+    mPSelfTlv->field_2_length = static_cast<s16>(mSizeOfT);
 
     copyFn(pSelfTlv /* dst */, pTlv /* src */);
